requirements.txt
BeautifulSoup4==4.9.3
virtualenv
Naked
tqdm
asyncio
asyncio-throttle